[wip] Test u32 conversion

+++ android/src/main/java/io/emurgo/rnhaskellshelley/HaskellShelleyModule.java
@@ -1781,9 +1781,9 @@ public class HaskellShelleyModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public final void transactionBuilderNew(String linearFee, String minimumUtxoVal, String poolDeposit, String keyDeposit, Integer maxValueSize, Integer maxTxSize, Promise promise) {
+    public final void transactionBuilderNew(String linearFee, String minimumUtxoVal, String poolDeposit, String keyDeposit, Double maxValueSize, Double maxTxSize, Promise promise) {
         Native.I
-                .transactionBuilderNew(new RPtr(linearFee), new RPtr(minimumUtxoVal), new RPtr(poolDeposit), new RPtr(keyDeposit), maxValueSize, maxTxSize)
+                .transactionBuilderNew(new RPtr(linearFee), new RPtr(minimumUtxoVal), new RPtr(poolDeposit), new RPtr(keyDeposit), maxValueSize.longValue(), maxTxSize.longValue())
                 .map(RPtr::toJs)
                 .pour(promise);
     }
